Improved readability

+++ cities_game/capital_city.py
@@ -4,5 +4,5 @@ from cities_game.city import City
 
 
 class Capital(City):
-    def __init__(self, people_amount: int, level: int, position: np.ndarray[int]) -> None:
+    def __init__(self, people_amount: int, level: int, position: np.ndarray[float]) -> None:
         super().__init__(people_amount, level, position)
